pyproject.toml
#[build-system]
#requires = ["setuptools"]
#build-backend = "setuptools.build_meta"

[project]
name = "wikipedia"
version = "0.0.1"
description = "Classification of Wikipedia articles using GNN"
authors = [
  { name = "Rasmus Fiehn", email = "s204617@dtu.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "evidently>=0.4.36",
    "fastapi>=0.115.6",
    "gcloud>=0.18.3",
    "google-cloud>=0.34.0",
    "google-cloud-aiplatform>=1.78.0",
    "google-cloud-secret-manager>=2.22.1",
    "google-cloud-storage>=2.19.0",
    "httpx>=0.28.1",
    "hydra-core>=1.3.2",
    "invoke>=2.2.0",
    "locust>=2.32.6",
    "matplotlib>=3.10.0",
    "numpy>=2.2.1",
    "pandas>=2.2.3",
    "protobuf>=5.29.3",
    "pytest-cov>=6.0.0",
    "python-dotenv>=1.0.1",
    "pytorch-lightning>=2.5.0.post0",
    "torch>=2.5.1",
    "torch-geometric>=2.6.1",
    "typer>=0.15.1",
    "uvicorn>=0.34.0",
    "wandb>=0.19.3",
]

[tool.ruff]
line-length = 120
lint.select = ["I"]

[tool.coverage.run]
omit = ["tests/*"]

[dependency-groups]
dev = [
    "coverage>=7.6.10",
    "evidently>=0.4.36",
    "jupyter>=1.1.1",
    "loguru>=0.7.3",
    "markdown>=3.7",
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "ruff>=0.9.2",
]
